Flatten cache abstraction: inline BaseDataFrameCache into APICache
BaseDataFrameCache was an abstract base class with exactly one implementation
(APICache). Merged all logic into APICache directly and reduced base_cache.py
to a backward-compatible re-export alias. Removes unnecessary inheritance
indirection.
