+++ sbir_etl/utils/cache/api_cache.py
@@ -1,21 +1,21 @@
-"""Generic API response cache utility.
+"""File-based DataFrame cache for API responses.
 
-This module provides a unified file-based caching mechanism for API responses,
-replacing the separate PatentsViewCache and USAspendingCache implementations.
-It supports flexible cache keys, metadata, and expiration.
+Provides caching of API responses as parquet files with TTL-based expiration,
+metadata tracking, and cache management (clear, stats).
 """
 
+import hashlib
 import json
+from datetime import datetime, timedelta
 from pathlib import Path
 from typing import Any
 
+import pandas as pd
 from loguru import logger
 
-from .base_cache import BaseDataFrameCache
 
-
-class APICache(BaseDataFrameCache):
-    """Generic file-based cache for API responses."""
+class APICache:
+    """File-based cache for API responses stored as DataFrames."""
 
     def __init__(
         self,
@@ -32,20 +32,133 @@ class APICache(BaseDataFrameCache):
             enabled: Whether caching is enabled
             ttl_hours: Time-to-live for cache entries in hours (default: 24)
         """
-        super().__init__(cache_dir=cache_dir, enabled=enabled, ttl_hours=ttl_hours)
-        self._default_cache_type_value = default_cache_type
+        self.cache_dir = Path(cache_dir)
+        self.enabled = enabled
+        self.ttl_hours = ttl_hours
+        self.default_cache_type = default_cache_type
 
-    def _get_default_cache_type(self) -> str:
-        """Get the default cache type.
+        if self.enabled:
+            from sbir_etl.utils.path_utils import ensure_dir
+
+            ensure_dir(self.cache_dir)
+            logger.debug(
+                f"APICache initialized at {self.cache_dir} (TTL: {ttl_hours}h)"
+            )
+
+    def _generate_cache_key(
+        self,
+        uei: str | None = None,
+        duns: str | None = None,
+        company_name: str | None = None,
+        cache_type: str = "default",
+    ) -> str:
+        """Generate a cache key from company identifiers.
+
+        Args:
+            uei: Company UEI
+            duns: Company DUNS number
+            company_name: Company name
+            cache_type: Type of cache entry
 
         Returns:
-            Default cache type string
+            Cache key (SHA256 hash of normalized identifiers)
         """
-        return self._default_cache_type_value
+        parts = []
+        if uei:
+            parts.append(f"uei:{uei.strip().upper()}")
+        if duns:
+            parts.append(f"duns:{str(duns).strip()}")
+        if company_name:
+            normalized_name = " ".join(company_name.strip().upper().split())
+            parts.append(f"name:{normalized_name}")
+
+        if not parts:
+            raise ValueError("At least one identifier (UEI, DUNS, or name) must be provided")
+
+        parts.append(f"type:{cache_type}")
+        key_string = "|".join(sorted(parts))
+        return hashlib.sha256(key_string.encode("utf-8")).hexdigest()
+
+    def _get_cache_path(self, cache_key: str) -> Path:
+        """Get the file path for a cache key."""
+        return self.cache_dir / f"{cache_key}.parquet"
+
+    def _get_metadata_path(self, cache_key: str) -> Path:
+        """Get the metadata file path for a cache key."""
+        return self.cache_dir / f"{cache_key}.meta.json"
+
+    def _is_expired(self, metadata: dict[str, Any]) -> bool:
+        """Check if a cache entry is expired."""
+        if "cached_at" not in metadata:
+            return True
+
+        cached_at = datetime.fromisoformat(metadata["cached_at"])
+        expiration_time = cached_at + timedelta(hours=self.ttl_hours)
+        return datetime.now() > expiration_time
+
+    def get(
+        self,
+        uei: str | None = None,
+        duns: str | None = None,
+        company_name: str | None = None,
+        cache_type: str | None = None,
+    ) -> pd.DataFrame | None:
+        """Retrieve cached data if available and not expired.
+
+        Args:
+            uei: Company UEI
+            duns: Company DUNS number
+            company_name: Company name
+            cache_type: Type of cache entry (uses default if not provided)
+
+        Returns:
+            Cached DataFrame if found and valid, None otherwise
+        """
+        if not self.enabled:
+            return None
+
+        if cache_type is None:
+            cache_type = self.default_cache_type
+
+        try:
+            cache_key = self._generate_cache_key(
+                uei=uei, duns=duns, company_name=company_name, cache_type=cache_type
+            )
+            cache_path = self._get_cache_path(cache_key)
+            metadata_path = self._get_metadata_path(cache_key)
+
+            if not cache_path.exists() or not metadata_path.exists():
+                return None
+
+            with open(metadata_path) as f:
+                metadata = json.load(f)
+
+            if self._is_expired(metadata):
+                logger.debug(
+                    f"Cache entry expired for {cache_key[:8]}... (cached at {metadata.get('cached_at')})"
+                )
+                cache_path.unlink(missing_ok=True)
+                metadata_path.unlink(missing_ok=True)
+                return None
+
+            df = pd.read_parquet(cache_path)
+            logger.debug(
+                f"Cache hit for {cache_key[:8]}... ({len(df)} rows, cached at {metadata.get('cached_at')})"
+            )
+            return df
+
+        except Exception as e:
+            logger.warning(f"Error reading cache: {e}")
+            try:
+                cache_path.unlink(missing_ok=True)
+                metadata_path.unlink(missing_ok=True)
+            except Exception:
+                pass
+            return None
 
     def set(
         self,
-        df,
+        df: pd.DataFrame,
         uei: str | None = None,
         duns: str | None = None,
         company_name: str | None = None,
@@ -59,25 +172,45 @@ class APICache(BaseDataFrameCache):
             uei: Company UEI
             duns: Company DUNS number
             company_name: Company name
-            cache_type: Type of cache entry
+            cache_type: Type of cache entry (uses default if not provided)
             **metadata: Additional metadata to store
         """
+        if not self.enabled:
+            return
+
         if cache_type is None:
-            cache_type = self._get_default_cache_type()
+            cache_type = self.default_cache_type
 
         # Remove identifiers from metadata if present to avoid conflict
         metadata.pop("uei", None)
         metadata.pop("duns", None)
         metadata.pop("company_name", None)
 
-        super().set(
-            df=df,
-            uei=uei,
-            duns=duns,
-            company_name=company_name,
-            cache_type=cache_type,
-            **metadata,
-        )
+        try:
+            cache_key = self._generate_cache_key(
+                uei=uei, duns=duns, company_name=company_name, cache_type=cache_type
+            )
+            cache_path = self._get_cache_path(cache_key)
+            metadata_path = self._get_metadata_path(cache_key)
+
+            from sbir_etl.utils.data.file_io import save_dataframe_parquet
+
+            save_dataframe_parquet(df, cache_path, index=False)
+
+            metadata_dict = {
+                "cached_at": datetime.now().isoformat(),
+                "cache_key": cache_key,
+                "cache_type": cache_type,
+                "row_count": len(df),
+                **metadata,
+            }
+            with open(metadata_path, "w") as f:
+                json.dump(metadata_dict, f, indent=2)
+
+            logger.debug(f"Cached {len(df)} rows for {cache_key[:8]}...")
+
+        except Exception as e:
+            logger.warning(f"Error writing cache: {e}")
 
     def clear(self, cache_type: str | None = None, expired_only: bool = False) -> int:
         """Clear cache entries.
@@ -98,7 +231,6 @@ class APICache(BaseDataFrameCache):
             metadata_path = self._get_metadata_path(cache_key)
 
             if not metadata_path.exists():
-                # Orphaned cache file or no metadata, safe to delete if we are not strict
                 if not expired_only:
                     cache_file.unlink(missing_ok=True)
                     cleared_count += 1
@@ -108,22 +240,18 @@ class APICache(BaseDataFrameCache):
                 with open(metadata_path) as f:
                     metadata = json.load(f)
             except Exception:
-                # Corrupted metadata
                 if not expired_only:
                     cache_file.unlink(missing_ok=True)
                     metadata_path.unlink(missing_ok=True)
                     cleared_count += 1
                 continue
 
-            # Filter by cache type if specified
             if cache_type and metadata.get("cache_type") != cache_type:
                 continue
 
-            # Check expiration if requested
             if expired_only and not self._is_expired(metadata):
                 continue
 
-            # Delete files
             cache_file.unlink(missing_ok=True)
             metadata_path.unlink(missing_ok=True)
             cleared_count += 1
@@ -171,7 +299,6 @@ class APICache(BaseDataFrameCache):
                 except Exception:
                     is_expired = True
             else:
-                # No metadata = effectively expired/invalid
                 is_expired = True
 
             if is_expired:
@@ -185,5 +312,9 @@ class APICache(BaseDataFrameCache):
             "total_size_mb": round(total_size / (1024 * 1024), 2),
             "cache_dir": str(self.cache_dir),
             "ttl_hours": self.ttl_hours,
-            "default_type": self._default_cache_type_value,
+            "default_type": self.default_cache_type,
         }
+
+
+# Backward compatibility alias
+BaseDataFrameCache = APICache
